Refactored the KSPTime classes
There is no longer a `elapsedYDHMS` property.
Instead, a `displayYDHMS` represents the date to actually
dispay (in UT mode). For specific elapsed time calculation,
this is done directly inside `stringYDHMS` on EMT mode, because
this is the only place where it's used.

Some struggle when modifying the departure and arrival
dates has also been modified by making the auto validation
less harsh in `TimeSelector`. Still some annoying
effects on the system time selector.